Moved quote handling unit tests into a dedicated class, CommandTreeTokenizerTests
